+++ pwa/src/App.js
@@ -1,45 +1,12 @@
-import React, { useState, useEffect } from 'react';
+import React, { useState } from 'react';
 
 import './App.css';
-import { onHwData } from './api.js';
-import Dashboard from './Dashboard.js';
-
-const initialState = {
-  fuelData: {
-    remainingPercentage: 0,
-    remainingLiters: 0,
-    remainingLitersEMA: 0,
-  },
-  gpsData: {
-    latitude: 0,
-    longitude: 0,
-    speed: 0,
-    trackAngle: 0,
-  },
-};
+import DriverDisplay from './DriverDisplay/DriverDisplay.js';
 
 function App() {
-  const [data, setData] = useState(initialState);
-
-  useEffect(
-    () => {
-      const unsubscribe = onHwData((data) => {
-        if (data.type === 'fuel_data') {
-          console.log('Updating fuel data', data.value);
-          setData(prevState => ({ ...prevState, fuelData: data.value }));
-        } else if (data.type === 'gps_data') {
-          console.log('Updating GPS data', data.value);
-          setData(prevState => ({ ...prevState, gpsData: data.value }));
-        }
-      });
-      return unsubscribe; // call unsubscribe when unmounting
-    },
-    [], // only run on mount and unmount
-  );
-
   return (
     <div className="App">
-      <Dashboard data={data} />
+      <DriverDisplay />
     </div>
   );
 }
