+++ pwa/src/DriverDisplay/DriverDisplay.css
@@ -1,0 +1,3 @@
+.DriverDisplay {
+  text-align: center;
+}
